+++ ja2lib/Strategic/PreBattleInterface.c
@@ -1714,7 +1714,7 @@ BOOLEAN PlayerGroupInvolvedInThisCombat(struct GROUP *pGroup) {
   return (FALSE);
 }
 
-BOOLEAN CurrentBattleSectorIs(INT16 sSectorX, INT16 sSectorY, INT16 sSectorZ) {
+BOOLEAN CurrentBattleSectorIs(u8 sSectorX, u8 sSectorY, INT16 sSectorZ) {
   INT16 sBattleSectorX, sBattleSectorY, sBattleSectorZ;
   BOOLEAN fSuccess;
